Stop parallel scans slamming resources
Currently when file change notifications are emitted a scan is performed
per every change. This can be troublesome especially during build processes
where potentially hundreds of files may change. The fix here is to ensure only
one scan is operating at a time.

We also are opening potentially thousands of files in parallel when running scans.
Instead we want to open them synchonously ensuring that they are closed and we don't
hog system resources for file opens. This could be optimized in the future to open a
small number in parallel (say 10) if the performance is needed here.